backend/pyproject.toml
[project]
name = "prelegal-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "litellm>=1.83.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.32",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
pythonpath = ["."]
